Reworking the Slider
Added more functionality to the Slider by exposing more properties and
subparts:

- The notion of "secondary value" to emulate buffering in the slider, this
  is a common enough case.

- ValueTrack and SecondaryValueTrack subparts, to easy customize how the
  current value and the "buffering" value is shown.

- ValueIndicator subpart (plus its customization properties), that shows
  the current value. The actual text shown can be formatted by implementing
  the formatValue(v) function in the Slider.

- Animation support in the Slider itself using a Behavior (possible
  due to the separation of handle and fakeHandle in the
  implementation).

There is a trade off between the feature set ready available and the
flexibility. Since the valueTrack/secondaryTrack/valueIndicator are all
optional (and can be set to undefined without a problem), those features
are opt-in.

If one desires, can implement things "by hand" inside the groove and
the handle with information from the range model.

I still didn't dig further on what properties the "style developers"
will need available, we can export them as we see fit.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>

Merge-request: 42
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@nokia.com>