Adjusted the new Slider implementation
- Remove valueTrack and secondaryValue valueTrack
I am not sure we should add such extensive functionality
yet. This looks a bit too meego specific for now. Instead I
draw the valueTrack as part of the background and export
positionForValue to the loaded component.

- Position range within leftMargin and rightMargin
to allove "boxed" handles. We should not assume that
you allways want to clip the range within handlewidth/2