SplitterRow: remove splitterHandle property
This property turns out to be not a such a neat idea
after all. The reason is that SplitterRow cannot assign
it a value before the component is completed, and especially,
after all the child items has completed. This is late, and can
cause strange artifacts if the child items use the property
in bindings, and even worse, inside the onCompleted function. Also,
not all children have a splitter handle associtated.