Introduce a simple API for accessing configuration from QML
The application instantiates a Settings element and adds properties to
it. These properties can alias other fields in the graph or be bound to
externally -- like regular QML properties.

    Item {
        id: main
        height: 480
        width: 640
        color: settings.color

        Settings {
            id: settings
            property string color: "blue"
            property alias height: main.height
            property alias width: main.width
        }
    }

Behind the scenes, the element will grab saved state when loaded and
save it back as the application state changes.

* Implementation notes

This is a proof of concept.

The initial implementation is using QSettings and is sub-optimal. At the
very least, a QSettings object could be shared across all Settings
instances and I/O should be kept outside the main thread.

It is also worth noting that QSettings does not support notifications
for dynamically updating the UI. A different backend could.

* Bugs/Issues

Data is not saved in the destructor because I was gettings width and
height properties set to 0 when the window closes.

Currently using "group" property to map to QSettings group. Applications
can't use a "group" property.

Merge-request: 11
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@nokia.com>