Change the QCursorArea to set the ArrowCursor explicitly.
Prior to this fix it was not possible to use a QCursorArea
with an arrow cursor that covers another QCursorArea with
some special cursor set. A typical use case would be
a popup that should never show any cursors of underlying items.

It is now also possible to reset the cursor, which actually
makes the QCursorArea do nothing (apart from using resources).

Merge-request: 17
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@nokia.com>