Port QWindowItem to Qt 5.
Add ifdefs to the Qt 4 version. Depending on what
branching strategy we chose we either merge this back
to the master branch or strip out the Qt 4 code later
on.

Add QWindowWidget for embedding QWindows in QWidget
hierarchies.

QTopLevelWindow is still a QMainWindow (we want to
re-use the menu implementation), but the center
widget is now a QWidnowWidget with an embedded 
QQuickView.


Change-Id: I173878efc869d8224e7330a13b37e85730d8ecbf
Reviewed-by: Morten Johan Sørvig <morten.sorvig@nokia.com>