PageStack: add recursion guard
We need to protect the page stack while it updates properties to
ensure that the update is atomic. Otherwise the user can e.g
push a new page upon receiving Page.onStatusChanged or
Page.onCompleted, which will recursively start a new transition
while we prepare for another. Calling push/pop while the transition
runs is fine.

Change-Id: Ib9453b1a61f8c8cee19f5e748bb75a1a30614119
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@digia.com>