Tests: Refactor build and use of test plugin
The test plugin is now built as a standard plugin and
does not need to be installed. Using qml_plugin.prf
is not necessary any more and it didn't work particulary
well for our use case. With this change we remove the
"make install" hack in QMAKE_POST_LINK.

In qtdesktop test, use IMPORTPATH so qmlscene can
find the test plugin in a custom location when
make check is run. This is needed because qml test
files launched via qtdesktop should be able to import
the test plugin if needed (i.e. tst_tableview.qml).

Change-Id: I54d64bce5b1d5fbf1233a122d0c205f71193d55d
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>