qdoc: QtDesktop1::Menu is not correctly documented.
When documenting a QML component, each qdoc comment must immediately
precede the element being documented. In this file, that rule was
violated. The property and method comments should be removed, because
they don't apply to anything in the Menu. Or, if they should be there,
they should be reintroduced so that each one immediately precedes the
element being documented. The \qmltype comment is probably out of date
now as well.

Task-number: QTBUG-29740
Change-Id: Id8048e5f9d6d158305110a7c561224a8033ab088
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>