Example: Add copy/paste/cut actions to the gallery
Save the last item with focus to reuse while
using shortcuts at the application level.

Call the function (if it exists) on the last item
with focus when the related shortcut is triggered.

This way a shortcut works both when it is called
directly or when it is called from the menu.

Change-Id: I5fb26a8793fa07c75da6c737cff4701291e0c1bc
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>