MenuBar: Fix appearance on Windows, Ubuntu
On top of adding the pixels below the menubar, we also set the
proper font on the painter before rendering. This font was
originally set when initializing the QPainter from the widget
to be rendered. Since we don't have any widget, we set it just
before the style rendering calls.

Also added more pixel metrics properties.

Change-Id: I631699458403ecc3d83d5c9bbf57805d18547c71
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>