Menu: Remove 'model' property, ContextMenu
In the near future, we hope that QML types like Creator/Instantiator
(currently under review) will help simplify model binding, and object
creation and deletion. Also, given the differences between QML models
and QAbstractItemModel APIs, it's hard to provide a unified and
elegant solution for this right now.

ComboBox gets the part of the logic that ContextMenu was responsible
for, extended to support string list and number models. Auto-tests
updated and XFAIL removed where applicable.

Change-Id: I9f5d4059644c495bffff76fb7c353e6fe7fde62e
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@digia.com>