Rearrange only the topmost layout.
Since we also rearrange in geometryChanged(), this should ensure
that we recurse down and arrange all sublayouts.

This assumes that objects are constructed in bottom-up order.
(That's what componentComplete() indicates.)

Previously we would risk rearranging sublayouts several times:
1. Sublayout got componentComplete -> geometryChanged -> rearrange()
2. Parent layout got componentComplete -> geometryChanged -> rearrange()
3. Since the sublayout was a layout item of the parent layout it could
   get a geometry change again, causing it to rearrange again.

This also fixes the issue where the implicitWidth/Height of the layout
was not immediately set after componentComplete, which could cause it
to get the wrong initial size.

Change-Id: I63a35dd934cd3ace01fab6319333d531631a6f4e
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@digia.com>