ScrollView: Fix crash on exit
Reproducible with the gallery example application, Launch it, select
the TextField, and quit the application.

The following short example also reproduces the crash in the same way,
but selecting the TextInput instead.

import QtQuick 2.1
import QtQuick.Window 2.1
import QtQuick.Controls 1.0

Window {

    property bool blah: !!activeFocusItem && !!activeFocusItem["copy"]

    Column {
        TextInput {
            width: 200
            height: 30
        }
        TextArea {
            width: 200
            height: 200
        }
    }
}

Whether the real cause is in ScrollView or deeper in QtQuick is yet to
be throughly investigated. But, hey, that's one crash less right now.

Change-Id: Idf358775b8b0333261ecfec7a2fefd3c5b931cc8
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>