ComboBox: Set implicitWidth, text eliding
Getting the extra width from the arrows turns out to need
an extra StyleItem just for that. Leaving a 35-pixel margin
is more than enough for the most common styles. In any case
the Combobox itself should be wide enough to contain all the
texts, or if the texts are really long (file name, etc.)
eliding won't be pixel critical either.

Task-number: QTBUG-30377
Change-Id: Ifab56f87b518102522f879c560b7b45802509fa3
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>