TextField: Fix baseline distance to panel frame
Still not perfect, and both the cursor and selection height are
one pixel too tall on Mac (and possibly Windows). On Ubuntu, Gtk 2
looks almost identical to native. Fusion also looks almost identical
compared to QLineEdit.

The basic idea is to override (again) what the style does, since we
do the painting in the Qt Quick side.

Change-Id: I179c2dfe6c76dd35f8bb87a5cd7c183f8e0f67e5
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>