Fixed a bug where spans across empty cells got broken.
If a row/column is only used only because of the spanning of an item,
the cell should be treated as it didn't exist. We keep track of this
with the "ignore" bit array.

The old code would always start from the row/column at position 1.
In the attached testcase this made the effectiveRowSpan become larger
than actually needed.

Task-number: QTBUG-30255

Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>
Change-Id: I410ed373f408014333d3964e2ddbcfeb25bec3a6
(cherry picked from qtbase/4f072e2d3d7e429359ff15a615d02712bff7ee51)
Reviewed-by: Jan Arve Sæther <jan-arve.saether@digia.com>