Menu: Add API for inserting items, separators, submenus
In particular, Menu::insertItem() is now overloaded. It can take
a string as second parameter, in which case it will create a new MenuItem
and return it after inserting it, or it can take any QtMenuBase based class
pointer and just insert it. The latter use case is useful when populating
a menu with the Instantiator type.

Currently, addSeparator and insertSeparator don't return any value since
MenuSeparators are not configurable.

Change-Id: I9659ea0ee9d76ed5163316be1a5c7a8f03eaa28f
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>