Remove a function that doesn't serve any purpose anymore.
setupItemLayout was supposed to do initialization on each item before
it got added to the layout, but all it did was to add the attached
property to the item, which is really not needed anymore.
The layout code does not work under the assumption that each item has
the attached property anyway, so we can remove it. (And in addition
avoid some potential allocations)

Change-Id: I6d7e3e5b13531dfeaceef495de144dcf87d38cc4
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>