Make tab accessibility work.
The PageTab role is actually the button in the TabBar
you click to change tabs - it's not the Tab item
itself.

PageTabs are treated as leaf items with no children
by QML accessibility, which is obviously very wrong
for Tab container items.

There is currently no specific "TabContainer" role in
Qt accessibility. Use LayeredPane as a replacement.

Change-Id: I4728025aacd4b7c370c419533a25f58950f24b6c
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>