SplitView: always respect min/max
If the user sets Layout.maximumWidth on the fillWidth item, we now
respect it. This can potentially make the splitview look buggy, since
the fillWidth item might then be restrained from filling out remaining
space. The same with the height of each item; you can now control the
maximum height. This means that each item can end up with a different
height. The upshot is that if you add items to a splitview that
cannot grow, splitview will not try to resize it.

Change-Id: If1cc82aefa939898a3b159da6b006f95af42599f
Reviewed-by: Jan Arve Sæther <jan-arve.saether@digia.com>