TableView: Fix activated() not emitted
The mousearea in TableView has to accept mouse event in onPressed to
receive upcoming double click event.

Regression introduced by Iff3e5c1.

Instead of not accepting the event in onPressed, set the MouseArea
property propagateComposedEvents to true and does not accept the onClicked
event.

Also uses QStyle::SH_ItemView_ActivateItemOnSingleClick to check if the
table is activated on single or double-clicking. By default, it is activated
on double-clicking.

Change-Id: I744174d7b31efebfae5e8db4db7b99a544ecf50f
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>