Fixed some bugs in how effective size hints were calculated
Do not bound explicit preferred size with implicit sizes. This
means that if Layout.preferredWidth was *explicitly* set to 50,
and Layout.maximumWidth had the implicit value 20, the effective
maximum width would be expanded to 50 in order to not disregard
the explicitly set preferred width. (covered by the tag
"expandMaxToExplicitPref" in test_sizeHint)

Note that this doesn't break any autotests, but the row layout
autotest was slightly changed to be able to add the new test cases.

This should make the behavior match the behavior of
QGraphicsGridLayout and QGraphicsLinearLayout

Change-Id: Ia23c8ef909827f14349906c003c72bb83689ef9a
Reviewed-by: Caroline Chao <caroline.chao@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>