Menu: Support QQuickWindows embedded in QWidgets
We translate the menu popup window by its transient parent origin
coordinates, but if it's embedded in a QWidget, we would only translate
by the transient parent's origin inside that QWidget. So, in this case.
we translate by the transient parent's origin in the global coordinate
reference system.

Task-number: QTBUG-31659
Change-Id: I09aa64f2d61d227d4fcd6d6e22a25453e1f2dc76
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>