Fix missing tool tip on Button
This binding was trying to hide the tooltip when the tooltip
text is the same as the button text. Unforunatately the logic
was somewhat broken and could not even theoretically work
for the case when there was a tooltip text set on the
button itself. Since the user can locally fix this by simply
not setting a tool tip or setting it to an empty string, I
think it is better to simply get rid of the binding.

Change-Id: I80ad2941649915a4f0eb54d9ba1c590b0292662c
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>