SplitView: Always layout as if fillWidthItem is visible
Hiding the fillWidth item in the splitter is allowed, but
causes strange drawing/layout artifacts that can be
difficult to understand since the layout engine would ignore
all hidden items. This patch changes the implementation
by adding an exception to the rule. The layout engine
now treats the fillWidth item as if it is always visible.
Then, hiding it would just reveal a white field where the
fillItem would otherwise be, but the layout would otherwise
look correct.

Change-Id: I95ec51b595109e8f1efdc231c9a6697ee60aa483
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>