Allow mouse events when transient scroll bars are hidden
On mac, while transient scroll bars are not visible it was not
possible to interact with the viewport as the mouse events were
consumed by the scroll view.

This makes the ScrollBar mouse area disabled before a scroll is
triggered. Note that I also added a minor workaround to ensure
that we update the mouse cursor.

Task-number: QTBUG-32400
Change-Id: If091edf054fcab65feb9b8c8e5c45e25431f4be4
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>