Add qml/js files in resource
All the qml and js files are embedded in the Qt Quick Controls plugin
(for Controls, Styles and Private).

If ApplicationWindow.qml is missing, loading the files from resource
will be enabled. Otherwise, the files will be loaded from the local
path.

The purpose of this change is to make the deployment of Qt Quick
Controls easier while keeping the development tasks convenient. With
this solution:
- It is only necessary to deploy the qmldir and the plugin
qtquickcontrolsplugin found in the QtQuick/Controls folder.
- By default the files used are the ones found in the local folder.
Autocompletion and debugging capabilities remain unchanged.

Task-number: QTBUG-31565

Change-Id: I938ebe261c3c35f9e3b066d82c81accd9750edc9
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>