Fix broken testBench properties with v4 changes
There is a small behavior change in v4 that prevents the old code
from working as change signals are no longer prefixed with "on".

However as Simon pointed out, we can directly check
if an object is a function which was the whole point of the messy
filtering code so I have updated the test to do this.

Change-Id: I20ac0ca5cf5c602939aa1b679b78e275dec2f3e2
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>
Reviewed-by: Caroline Chao <caroline.chao@digia.com>