Add textAlignment to TableView
It is unclear if the name should be horizontalAligment or
textAlignment since we are inconsistent between TableViewColumn
and the styleData for the item delegate. But since we call it
textAligmnent in the item delegate I opted for that.

We also need a new convention for adding versions to styleData
groups.

Change-Id: I68b7355f663d2de0e712f849f9cad0a4910abecd
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>