Fix scroll bar behavior when clickToPosition is enabled
There were two issues with the scrollbar when using this behavior:

Issue 1 - The scroll bar did not account for the grooveRect position,
so the handle was not centered correctly around the mouse cursor.

Issue 2 - When click-and-holding outside the handle area, you could not
continue to move the handle afterwards.

Task-number: QTBUG-33133

Change-Id: I251f684f767ce5a372eb0d2b5e6c421f88ac7fbe
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>