Reduce chaos during deconstruction
When a layout is destroyed, it will during its destruction unparent its
children and possibly change their visibility. The visibility change
will recurse down to all descendants.
Because of these slots, this information will again be communicated up
to the layout, which again might cause the layout to adjust to this new
change _while_it_is_deconstructing_. This makes it chaotic, is a waste
of resources and makes it harder to maintain, and could in worst case
cause crashes.

Change-Id: I6b3111b8ecc4ec4312dd36068b4b4b4e6529b9ef
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>