SplitView: fix hidden item after fillWidth item bug
Some of the calculations in SplitView assumed that if
item with index i was hidden, then so was handle at
index i as well. This is not correct. The reason is
that a the handle at i belongs to either item i
or item i+1, depending on where the fillWidth item is.
And this caused a bug when hiding the item to the
right of the fillWidth item.

Conflicts:
	tests/auto/controls/data/tst_splitview.qml

Change-Id: I59967813f264a856181c57e32c4820d09d7eb363
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>