Don't apply baseline calculations if Qt.AlignBaseline is not set
Parts of the code assumed that if q_minimumDescent or q_minimumAscent
was valid, it would reserve enough space in each case for the combined
ascents and descents of all items in a row.

The extreme case was when one item had a baseline at its bottom edge,
and the next item had the baseline at its top edge, the layouts minimum
height tried to reserve space for such a case even though baseline
alignment was not desired.

The last case in the test demonstrates such a (valid) case. The problem
was that the same logic was applied to the first case.

Change-Id: Ie24503b1a5f7333f16ed84ebe01ab8d48becb4e2
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>