Expose contentArea as contentItem in ApplicationWindow
This will intentionally shadow the QQuickWindow::contentItem

This enables users of ApplicationWindow to correctly set the
{min,max}imumHeight for the contents of the ApplicationWindow
(excluding menubar, toolbar and statusbar).

With this information at hand, the ApplicationWindow can easily deduct
the correct total size of its {min,max}imumHeight *including* the
menubar, toolbar and statusbar.

Previously, it was for instance not possible to set an exact
minimumHeight because it didn't take the toolbar, statusbar or
menubar into consideration

One small disadvantage is that the contentItem.minimumWidth
cannot be autocompleted. qmlplugindump does not inspect the members
of ContentItem.

Task-number: QTBUG-32443
Change-Id: I9903e7d4b5e6db7bd7c3c46351a2bb5578f67a57
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>