Introducing styling for Menu, MenuBar
MenuStyle

We expose the frame and item properties together with some
convenience properties (basically color related).

The menu item data is exposed via a styleData object, similarly to
what's done in TabViewStyle.

In addition, we introduce MenuStyle.menuItem which brings some
convenience when it comes to overriding subcontrols of the menu item.

MenuBarStyle

We expose background and menuBarItem. The menubar item's properties
are accessible through the styleData property in scope.

Style cascading

Additionally, MenuBarStyle has a menuStyle property that will apply
to all its menus and their submenus. Similarly, assigning a style
to a Menu object, will apply it to its submenus. It's still possible
to override the parent menu's style by declaring its own.

[ChangeLog][QtQuickControls][Styles]Menu and MenuBar are now styleable
Change-Id: Ib724b7a6426bdfce5da314620d51dfaa76f76500
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>