Make sure menu bar has a parent window
On Android we depend on the parent window of the menu bar to
know when to show which menu bar. The menu bar parent was
set in the onVisibleChanged signal of the background item, but
this visible property is never changed. Adding a signal handler
for the visible property of the window as well fixes the issue.

Task-number: QTBUG-34681
Change-Id: I2cbafabe165420fd1697fbe51a8f393eb5158ffd
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>