Prepare to move the gridlayout engine so it can be shared.
Currently, the gridlayout engine is duplicated in QtQuick.Layouts module
and in QtWidgets module.
The plan is to move the gridlayout engine from QtWidgets to QtGui and
export it privately from there. However, that would produce linkage
errors when linking the QtQuick.Layouts module.
As a temporary solution we put the gridlayout engine in a namespace
until the move has been completed.

Change-Id: I8b43583cadff8846b5ed275f7fb5614737d903ad
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@digia.com>