ApplicationWindow has the WindowFullscreenButtonHint flag by default
It should be possible to go fullscreen by default in most applications,
unless there is a reason not to, in which case the developer can turn
this flag back off.  It's also not so intuitive to turn it on, because
one needs to remember to specify Qt.Window as well, to avoid losing
the toolbar.  So making it the default avoids some mistakes/surprises.
Furthermore QWidget has this flag by default.

Task-number: QTBUG-33607
Change-Id: I350ffb38459d2a305353efe0db29335fe354a441
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>
Reviewed-by: Liang Qi <liang.qi@digia.com>