Fix desktop style animations to stop when the control is hidden
QStyleAnimation automatically stopped for hidden QWidgets, but didn't
know anything about QQuickItems and kept animating regardless of their
visibility. QStyleAnimation was changed so that it will keep animating
only as long as the animation target accepts animation updates. This
change ensures that the style animation updates are accepted only when
the control is visible.

Note: In order to restart such style animations, QQuickStyleItem needs
to be manually updated when it becomes visible.

Task-number: QTBUG-35319
Change-Id: Ia053912883b9babb184a5467adfc90dd619b3dc5
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>