Make StatusBar and ToolBar focus scopes
The problem is that the content item of both these controls have
the focus property set to true. This property will then affect the
implicit focus logic and should be protected by a focusscope.

Currently adding a ToolBar or StatusBar to an application could
result in this content item grabbing focus implicitly even if the
application only sets focus: true on a single item.

Task-number: QTBUG-34713
Change-Id: I94cb79d25e4cfd727915f5b5ab612ad5c0132db7
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>