FileDialog: new features in DefaultFileDialog.qml
Use a ComboBox for the filters.  Add a sidebar showing shortcuts for the
common paths and drives that the user will most likely need, and an
editable list of favorite paths.

[ChangeLog][QtQuickDialogs][FileDialog] The FileDialog fallback QML
implementation now uses QtQuick.Controls, and has added features
including a Combobox for file extension filters, and a sidebar with
common paths, drives/volumes and favorite paths.

Change-Id: I228ebdadcf338917db81a2b3d03d3c0b09584e7d
Reviewed-by: Liang Qi <liang.qi@digia.com>