Fix implicit width for SpinBox
This patch ensures that we take the minimum width of the SpinBox
into account when getting the size from the style item.

The patch still relies on the height of the TextField since we
want to ensure that they keep similar height and the existing
spinbox code behaves somewhat erratic.

Task-number: QTBUG-36283
Change-Id: I50d14e5afb2576548c381e7f70063dfcb727e7d6
[ChangeLog][QtQuickControls]Fixed implicit width for SpinBox
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>