Add currentFont property to FontDialog.
QFontDialog has this property. This patch effectively means that
the font property will be set when the dialog closes, instead of
whenever the current font in the dialog changes, so pressing cancel
will revert the font to what it was before the dialog was opened.

Align with e73d503b84be12e7bf28491356b459ca309e7c95 in qtdelarative,
the similar change for color dialog.

[ChangeLog][QtQuickControls][FontDialog] Added currentFont property.

Task-number: QTBUG-36574
Change-Id: I4140966543a677d8e654dbaac170a2fe8d3579a1
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>