Avoid warning when maximum size < minimum size
This only happens in intermediate states, where both *minimumHeight*
and *maximumHeight* is supposed to be updated.
However, since the order of the bindings is not know, we might
have intermediate states where minimumHeight > maximumHeight.

In our case minimumHeight was set first to a smaller size than
maximumHeight, this would trigger the height binding, causing it to try
to set its geometry while its minimumHeight and maximumHeight properties
were not in sync.

This also happened when maximumWidth < minimumWidth.

The output was something like this:

qwindowswindow.cpp(1306):QWindowsWindow::setGeometry: Attempt to set a
size (116x190) violating the constraints(0x200 - 16777215x190) on
window ApplicationWindow_QMLTYPE_12_QML_51/''

Change-Id: Ia4f6e340b608bb4cff5a35bc887adc4eea7efbad
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>