Fixed offset for overlay in moveable TableViewColumns
The Loader needs to be reparented in order to draw the overlay on
top of everything. In addition we need to take both the item offset and
the contentX into account. Note I also tweaked the actual drop offset
by headerRowDelegate.width/2 as that feels a bit more balanced.

[ChangeLog][QtQuickControls][TableView] Fixed an issue with
the header overlay offset while re-arranging TableViewColumns

Task-number: QTBUG-37180
Change-Id: Ia80049616bb30e1391948fdeaf3f8d68afd7ce3e
Reviewed-by: Caroline Chao <caroline.chao@digia.com>
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>