Slider: don't prevent stealing until drag threshold is exceeded
If a Slider is placed in a Flickable (or ListView, etc.),
preventStealing means that you cannot flick the flickable by dragging
over the Slider.  It's better to postpone setting that until we
detect whether the user is really attempting to drag the slider,
because we don't want to rule out use cases such as vertical sliders
in a horizontal ListView or vice-versa, so that you drag in
one direction to move the slider and in another direction to flick.
Platform drag threshold is exposed in Settings (from QStyleHints).
(reverts/refines c9665f3e2d1eb12f1cc5a569c20f5d569fafde7f)

Change-Id: I82feda6028f6effa5b26ad575a1933e48e971453
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@digia.com>