Try to maintain bindings to checked property in CheckBox
It is currently impossible to keep a binding to the checked property
because any external manipulation of checked will cause it to be
internally overwritten based on the implicitly triggered
checkedStateChanged signal. This adds a protection mechanism similar
to what we already have on the checkedStateChanged signal.

[ChangeLog][QtQuickControls][CheckBox]Fixed a problem incorrectly causing
bindings to the checked state to break.

Task-number: QTBUG-31627
Change-Id: I506fa34561529a414c7b088001f96e8f283edd89
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>