Disable mipmapping on the style item
The recent removal of QSGPlainTexture::setHasMipMaps() made the texture
material enable mipmapping on the texture by default. Previously this
was not the case, the material did not enable mipmapping on the texture
regardless of the value of mipmapFiltering.

The default value (Nearest) now properly enables mipmapping but this is
not ideal for the style item. Therefore we disable mipmapping explicitly.

Task-number: QTBUG-37690
Change-Id: I15607fd0c3a06f0cefc416b74f6cee25ef7966bd
Reviewed-by: Gunnar Sletta <gunnar.sletta@jollamobile.com>