Implement input for entering non-standard font size.
User should be able to enter non-standard font size which is not
available in the list of standard font sizes. The selected value
in the list of standard font sizes should synchronize with the
input field for the non-standard point size value, and vice versa
if there is a match.

Change-Id: I63b855054b2feb4dad339fe8ec06be5179bb2a37
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>