Dialogs: modality applies to fake window decorations too
For the sake of consistency with dialog behavior on desktop platforms,
when a modal dialog is shown on a display that doesn't support windows,
and therefore fake window decorations are used to make the dialog look
like a Window while actually being an Item in the same scene graph,
you can't dismiss it by tapping outside.  If it's non-modal you can.

[ChangeLog][QtQuick][Dialogs]when a modal dialog is shown on a display
that doesn't support windows, you can't dismiss it by tapping outside

Task-number: QTBUG-37860
Change-Id: I1150f53774f5d70e88e2b2f5a9d3524a8a00dbec
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>